MAK COFFEE
MAK COFFEE is a website that contains information about a cafe called MAK COFFEE such as an introduction to the shop, menus containing names of dishes, pictures, prices linked via firebase so that data can be updated at any time. , information and how to contact the owner. In addition, the website also has the function of measuring the current ambient temperature.